+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * Running last test command is now saved per window #19
 
+### Changes
+
+* Removed command `phpunit_command` #31
+
 ## 0.4.0
 
 ### New Features